- split off class BlkFilesystem from class Filesystem
