+++ storage/Filesystems/Filesystem.h
@@ -1,6 +1,6 @@
 /*
  * Copyright (c) [2014-2015] Novell, Inc.
- * Copyright (c) 2016 SUSE LLC
+ * Copyright (c) [2016-2017] SUSE LLC
  *
  * All Rights Reserved.
  *
@@ -34,9 +34,6 @@
 namespace storage
 {
 
-    class BlkDevice;
-    class ContentInfo;
-
 
     enum class FsType {
 	UNKNOWN, REISERFS, EXT2, EXT3, EXT4, BTRFS, VFAT, XFS, JFS, HFS, NTFS,
@@ -64,17 +61,6 @@ namespace storage
 
 	FsType get_type() const;
 
-	virtual bool supports_label() const = 0;
-	virtual unsigned int max_labelsize() const = 0;
-
-	const std::string& get_label() const;
-	void set_label(const std::string& label);
-
-	virtual bool supports_uuid() const = 0;
-
-	const std::string& get_uuid() const;
-	void set_uuid(const std::string& uuid);
-
 	const std::vector<std::string>& get_mountpoints() const;
 	void set_mountpoints(const std::vector<std::string>& mountpoints);
 	void add_mountpoint(const std::string& mountpoint);
@@ -85,32 +71,8 @@ namespace storage
 	const std::list<std::string>& get_fstab_options() const;
 	void set_fstab_options(const std::list<std::string>& fstab_options);
 
-	const std::string& get_mkfs_options() const;
-	void set_mkfs_options(const std::string& mkfs_options);
-
-	const std::string& get_tune_options() const;
-	void set_tune_options(const std::string& tune_options);
-
-	/**
-	 * Set the ResizeInfo. Only use for testsuites.
-	 */
-	void set_resize_info(const ResizeInfo& resize_info);
-
-	ContentInfo detect_content_info() const;
-
-	/**
-	 * Set the ResizeInfo. Only use for testsuites.
-	 */
-	void set_content_info(const ContentInfo& content_info);
-
-	static std::vector<Filesystem*> find_by_label(const Devicegraph* devicegraph,
-						      const std::string& label);
-
 	static std::vector<Filesystem*> find_by_mountpoint(const Devicegraph* devicegraph,
 							   const std::string& mountpoint);
-
-	// TODO class BlkFilesystem for not nfs, tmpfs?
-	std::vector<const BlkDevice*> get_blk_devices() const;
 
     public:
