+++ storage/Filesystems/UdfImpl.cc
@@ -34,7 +34,7 @@ namespace storage
 
 
     Udf::Impl::Impl(const xmlNode* node)
-	: Filesystem::Impl(node)
+	: BlkFilesystem::Impl(node)
     {
     }
